+++ go.work
@@ -1,8 +1,8 @@
 go 1.20
 
 use (
-	./cli
+	./lcli
 	./config
 	./lib
-	./ui
+	./lcui
 )
